Rework ansible plugin files for migration
