+++ .github/workflows/gradle.yml
@@ -1,0 +1,34 @@
+name: Java CI
+
+on: [push]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+      - name: Get Fetch Tags
+        run: git -c protocol.version=2 fetch --tags --progress --no-recurse-submodules origin
+        if: "!contains(github.ref, 'refs/tags')"
+      - name: Set up JDK 1.8
+        uses: actions/setup-java@v1
+        with:
+          java-version: 1.8
+      - name: Grant execute permission for gradlew
+        run: chmod +x gradlew
+      - name: Build with Gradle
+        run: ./gradlew build
+      - name: Get Release Version
+        id: get_version
+        run: VERSION=$(./gradlew currentVersion -q -Prelease.quiet) && echo ::set-output name=VERSION::$VERSION
+      - name: Upload azure plugin jar
+        uses: actions/upload-artifact@v1.0.0
+        with:
+          # Artifact name
+          name: Grails-Plugin-${{ steps.get_version.outputs.VERSION }}
+          # Directory containing files to upload
+          path: build/libs/ansible-plugin-${{ steps.get_version.outputs.VERSION }}.jar
