+++ .gitignore
@@ -75,3 +75,6 @@ hs_err_pid*
 
 # Avoid ignoring Gradle wrapper jar file (.jar files are usually ignored)
 !gradle/wrapper/gradle-wrapper.jar
+
+# OS Files
+.DS_Store
