+++ deploy-docker.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
 ./gradlew jar
-docker build --pull -t batix/rundeck-ansible:latest -t batix/rundeck-ansible:${TRAVIS_TAG} .
+docker build --pull -t rundeckpro/ansible-plugin:latest -t rundeckpro/ansible-plugin:${TRAVIS_TAG} .
 docker login -u="${DOCKER_USERNAME}" -p="${DOCKER_PASSWORD}"
-docker push batix/rundeck-ansible
+docker push rundeckpro/ansible-plugin
