+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'ansible-plugin'
+rootProject.name = 'rundeck-ansible-plugin'
