+++ src/main/groovy/com/rundeck/plugins/ansible/ansible/AnsibleException.java
@@ -1,4 +1,4 @@
-package com.batix.rundeck.core;
+package com.rundeck.plugins.ansible.ansible;
 
 import com.dtolabs.rundeck.core.execution.workflow.steps.FailureReason;
 import com.dtolabs.rundeck.core.execution.workflow.steps.StepException;
