+++ src/main/groovy/com/rundeck/plugins/ansible/ansible/AnsibleInventory.java
@@ -1,4 +1,4 @@
-package com.batix.rundeck.core;
+package com.rundeck.plugins.ansible.ansible;
 
 import java.util.HashMap;
 import java.util.Map;
