+++ src/main/groovy/com/rundeck/plugins/ansible/ansible/AnsibleInventoryBuilder.java
@@ -1,4 +1,4 @@
-package com.batix.rundeck.core;
+package com.rundeck.plugins.ansible.ansible;
 
 import com.dtolabs.rundeck.core.common.INodeEntry;
 import com.dtolabs.rundeck.core.plugins.configuration.ConfigurationException;
