+++ src/main/groovy/com/rundeck/plugins/ansible/ansible/PropertyResolver.java
@@ -1,4 +1,4 @@
-package com.batix.rundeck.core;
+package com.rundeck.plugins.ansible.ansible;
 
 import java.util.Map;
