+++ src/main/groovy/com/rundeck/plugins/ansible/util/Listener.java
@@ -1,4 +1,4 @@
-package com.batix.rundeck.utils;
+package com.rundeck.plugins.ansible.util;
 
 public interface Listener {
